add integration test

+++ src/main/java/com/example/school/service/CourseService.java
@@ -2,6 +2,7 @@ package com.example.school.service;
 
 import com.example.school.repository.CourseRepository;
 import com.example.school.repository.model.Course;
+import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,11 +31,17 @@ public class CourseService {
         return courseRepository.save(course);
     }
 
+    @SneakyThrows
     public Course update(Course course) {
+        if(!courseRepository.existsById(course.getId()))
+            throw new Exception("Course doesn't exist");
         return courseRepository.save(course);
     }
 
+    @SneakyThrows
     public void delete(Long id) {
+        if(!courseRepository.existsById(id))
+            throw new Exception("Course doesn't exist");
         courseRepository.deleteById(id);
     }
 }
